src/Projeto_Cannoli/requirements.txt
pandas
dash
plotly
scikit-learn
supabase
python-dotenv
flask
gunicorn             # <--- NOVO
psycopg2-binary      # <--- NOVO (Para conexão PostgreSQL)